+++ src/modules/disk/option.h
@@ -31,4 +31,5 @@ typedef struct FFDiskOptions
     FFDiskVolumeType showTypes;
     FFDiskCalcType calcType;
     FFColorRangeConfig percent;
+    bool ignoreRemote;
 } FFDiskOptions;